moved tester code to another file
